+++ test/mitll/langtest/server/database/custom/AnalysisTest.java
@@ -34,7 +34,7 @@ public class AnalysisTest {
     Map<String,CommonExercise> idToEx = new HashMap<String,CommonExercise>();
     for (CommonExercise ex : exercises) idToEx.put(ex.getID(),ex);
 
-    Map<Long, Map<String, Result>> userToResults = war.getResultDAO().getUserToResults(true, war.getUserDAO());
+    Map<Integer, Map<String, Result>> userToResults = war.getResultDAO().getUserToResults(true, war.getUserDAO());
     Map<String, Result> stringResultMap = userToResults.get(71l);
 
   //  Collection<CorrectAndScore> copy = getFirstCorrectAndScore(resultsForUser);
